+++ mobile-app/App.tsx
@@ -29,7 +29,7 @@ import {
   LanguageProvider,
   useLanguage,
 } from "@shared-contexts/LanguageProvider";
-import { DomainProvider } from "@contexts/DomainProvider";
+import { DomainProvider } from "@contexts/DomainContext";
 import * as Localization from "expo-localization";
 import { useColorScheme } from "react-native";
 import { WalletPersistence } from "@api/wallet";
