+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/PortfolioScreen.tsx
@@ -72,7 +72,7 @@ import { BottomSheetHeader } from "@components/BottomSheetHeader";
 import * as SplashScreen from "expo-splash-screen";
 import { useLogger } from "@shared-contexts/NativeLoggingProvider";
 import { bottomTabDefaultRoutes } from "@screens/AppNavigator/constants/DefaultRoutes";
-import { useDomainContext, DomainType } from "@contexts/DomainProvider";
+import { useDomainContext, DomainType } from "@contexts/DomainContext";
 import { AddressSelectionButtonV2 } from "./components/AddressSelectionButtonV2";
 import { ActionButtons } from "./components/ActionButtons";
 import {
