+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/components/DFIBalanceCard.test.tsx
@@ -9,6 +9,8 @@ import { LoanVaultState } from "@defichain/whale-api-client/dist/api/loan";
 import { DFIBalanceCard } from "./DFIBalanceCard";
 
 jest.mock("../../../../../contexts/DisplayBalancesContext");
+jest.mock("@contexts/DomainProvider");
+
 jest.mock("@react-navigation/native", () => ({
   useNavigation: jest.fn(),
   useIsFocused: jest.fn(),
